day04

+++ heima-leadnews-model/src/main/java/com/heima/model/wemedia/pojos/WmNews.java
@@ -54,7 +54,7 @@ public class WmNews implements Serializable {
      3 多图文章
      */
     @TableField("type")
-    private Short type;
+    private Integer type;
 
     /**
      * 图文频道ID
@@ -115,7 +115,7 @@ public class WmNews implements Serializable {
     private String images;
 
     @TableField("enable")
-    private Short enable;
+    private Integer enable;
 
     //状态枚举类
     @Alias("WmNewsStatus")
